More refactoring. Bad channels shown on vertical scroll bar.
